backend/requirements.txt
# Gogobe Backend Requirements

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
psycopg2-binary
fastapi
uvicorn
requests
redis
beautifulsoup4
lxml

# PDF Processing (already installed)
pdfplumber==0.10.3
pandas==2.2.0
openpyxl==3.1.2

# Optional: for better performance
python-multipart==0.0.6

# Scraping Engine
scrapy>=2.11.0
scrapy-playwright>=0.0.33
playwright>=1.40.0

# AI Integration
openai>=1.0.0









